+++ test/unit/operation_test.js
@@ -115,7 +115,7 @@ describe('Operation', function() {
   });
 
   describe('.pathPayment()', function() {
-    it('creates a pathPaymentOp', function() {
+    it('creates a pathPaymentStrictReceiveOp', function() {
       var sendAsset = new StellarBase.Asset(
         'USD',
         'GDGU5OAPHNPU5UCLE5RDJHG7PXZFQYWKCFOEXSXNMR6KRQRI5T6XXCD7'
@@ -151,7 +151,7 @@ describe('Operation', function() {
         Buffer.from(xdr, 'hex')
       );
       var obj = StellarBase.Operation.fromXDRObject(operation);
-      expect(obj.type).to.be.equal('pathPayment');
+      expect(obj.type).to.be.equal('pathPaymentStrictReceive');
       expect(obj.sendAsset.equals(sendAsset)).to.be.true;
       expect(
         operation
@@ -226,6 +226,238 @@ describe('Operation', function() {
       };
       expect(() => StellarBase.Operation.pathPayment(opts)).to.throw(
         /destAmount argument must be of type String/
+      );
+    });
+  });
+
+  describe('.pathPaymentStrictReceive()', function() {
+    it('creates a pathPaymentStrictReceiveOp', function() {
+      var sendAsset = new StellarBase.Asset(
+        'USD',
+        'GDGU5OAPHNPU5UCLE5RDJHG7PXZFQYWKCFOEXSXNMR6KRQRI5T6XXCD7'
+      );
+      var sendMax = '3.0070000';
+      var destination =
+        'GCEZWKCA5VLDNRLN3RPRJMRZOX3Z6G5CHCGSNFHEYVXM3XOJMDS674JZ';
+      var destAsset = new StellarBase.Asset(
+        'USD',
+        'GDGU5OAPHNPU5UCLE5RDJHG7PXZFQYWKCFOEXSXNMR6KRQRI5T6XXCD7'
+      );
+      var destAmount = '3.1415000';
+      var path = [
+        new StellarBase.Asset(
+          'USD',
+          'GBBM6BKZPEHWYO3E3YKREDPQXMS4VK35YLNU7NFBRI26RAN7GI5POFBB'
+        ),
+        new StellarBase.Asset(
+          'EUR',
+          'GDTNXRLOJD2YEBPKK7KCMR7J33AAG5VZXHAJTHIG736D6LVEFLLLKPDL'
+        )
+      ];
+      let op = StellarBase.Operation.pathPaymentStrictReceive({
+        sendAsset,
+        sendMax,
+        destination,
+        destAsset,
+        destAmount,
+        path
+      });
+      var xdr = op.toXDR('hex');
+      var operation = StellarBase.xdr.Operation.fromXDR(
+        Buffer.from(xdr, 'hex')
+      );
+      var obj = StellarBase.Operation.fromXDRObject(operation);
+      expect(obj.type).to.be.equal('pathPaymentStrictReceive');
+      expect(obj.sendAsset.equals(sendAsset)).to.be.true;
+      expect(
+        operation
+          .body()
+          .value()
+          .sendMax()
+          .toString()
+      ).to.be.equal('30070000');
+      expect(obj.sendMax).to.be.equal(sendMax);
+      expect(obj.destination).to.be.equal(destination);
+      expect(obj.destAsset.equals(destAsset)).to.be.true;
+      expect(
+        operation
+          .body()
+          .value()
+          .destAmount()
+          .toString()
+      ).to.be.equal('31415000');
+      expect(obj.destAmount).to.be.equal(destAmount);
+      expect(obj.path[0].getCode()).to.be.equal('USD');
+      expect(obj.path[0].getIssuer()).to.be.equal(
+        'GBBM6BKZPEHWYO3E3YKREDPQXMS4VK35YLNU7NFBRI26RAN7GI5POFBB'
+      );
+      expect(obj.path[1].getCode()).to.be.equal('EUR');
+      expect(obj.path[1].getIssuer()).to.be.equal(
+        'GDTNXRLOJD2YEBPKK7KCMR7J33AAG5VZXHAJTHIG736D6LVEFLLLKPDL'
+      );
+    });
+
+    it('fails to create path payment operation with an invalid destination address', function() {
+      let opts = {
+        destination: 'GCEZW',
+        sendMax: '20',
+        destAmount: '50',
+        sendAsset: StellarBase.Asset.native(),
+        destAsset: new StellarBase.Asset(
+          'USD',
+          'GDGU5OAPHNPU5UCLE5RDJHG7PXZFQYWKCFOEXSXNMR6KRQRI5T6XXCD7'
+        )
+      };
+      expect(() =>
+        StellarBase.Operation.pathPaymentStrictReceive(opts)
+      ).to.throw(/destination is invalid/);
+    });
+
+    it('fails to create path payment operation with an invalid sendMax', function() {
+      let opts = {
+        destination: 'GCEZWKCA5VLDNRLN3RPRJMRZOX3Z6G5CHCGSNFHEYVXM3XOJMDS674JZ',
+        sendMax: 20,
+        destAmount: '50',
+        sendAsset: StellarBase.Asset.native(),
+        destAsset: new StellarBase.Asset(
+          'USD',
+          'GDGU5OAPHNPU5UCLE5RDJHG7PXZFQYWKCFOEXSXNMR6KRQRI5T6XXCD7'
+        )
+      };
+      expect(() =>
+        StellarBase.Operation.pathPaymentStrictReceive(opts)
+      ).to.throw(/sendMax argument must be of type String/);
+    });
+
+    it('fails to create path payment operation with an invalid destAmount', function() {
+      let opts = {
+        destination: 'GCEZWKCA5VLDNRLN3RPRJMRZOX3Z6G5CHCGSNFHEYVXM3XOJMDS674JZ',
+        sendMax: '20',
+        destAmount: 50,
+        sendAsset: StellarBase.Asset.native(),
+        destAsset: new StellarBase.Asset(
+          'USD',
+          'GDGU5OAPHNPU5UCLE5RDJHG7PXZFQYWKCFOEXSXNMR6KRQRI5T6XXCD7'
+        )
+      };
+      expect(() =>
+        StellarBase.Operation.pathPaymentStrictReceive(opts)
+      ).to.throw(/destAmount argument must be of type String/);
+    });
+  });
+
+  describe('.pathPaymentStrictSend()', function() {
+    it('creates a pathPaymentStrictSendOp', function() {
+      var sendAsset = new StellarBase.Asset(
+        'USD',
+        'GDGU5OAPHNPU5UCLE5RDJHG7PXZFQYWKCFOEXSXNMR6KRQRI5T6XXCD7'
+      );
+      var sendAmount = '3.0070000';
+      var destination =
+        'GCEZWKCA5VLDNRLN3RPRJMRZOX3Z6G5CHCGSNFHEYVXM3XOJMDS674JZ';
+      var destAsset = new StellarBase.Asset(
+        'USD',
+        'GDGU5OAPHNPU5UCLE5RDJHG7PXZFQYWKCFOEXSXNMR6KRQRI5T6XXCD7'
+      );
+      var destMin = '3.1415000';
+      var path = [
+        new StellarBase.Asset(
+          'USD',
+          'GBBM6BKZPEHWYO3E3YKREDPQXMS4VK35YLNU7NFBRI26RAN7GI5POFBB'
+        ),
+        new StellarBase.Asset(
+          'EUR',
+          'GDTNXRLOJD2YEBPKK7KCMR7J33AAG5VZXHAJTHIG736D6LVEFLLLKPDL'
+        )
+      ];
+      let op = StellarBase.Operation.pathPaymentStrictSend({
+        sendAsset,
+        sendAmount,
+        destination,
+        destAsset,
+        destMin,
+        path
+      });
+      var xdr = op.toXDR('hex');
+      var operation = StellarBase.xdr.Operation.fromXDR(
+        Buffer.from(xdr, 'hex')
+      );
+      var obj = StellarBase.Operation.fromXDRObject(operation);
+      expect(obj.type).to.be.equal('pathPaymentStrictSend');
+      expect(obj.sendAsset.equals(sendAsset)).to.be.true;
+      expect(
+        operation
+          .body()
+          .value()
+          .sendAmount()
+          .toString()
+      ).to.be.equal('30070000');
+      expect(obj.sendAmount).to.be.equal(sendAmount);
+      expect(obj.destination).to.be.equal(destination);
+      expect(obj.destAsset.equals(destAsset)).to.be.true;
+      expect(
+        operation
+          .body()
+          .value()
+          .destMin()
+          .toString()
+      ).to.be.equal('31415000');
+      expect(obj.destMin).to.be.equal(destMin);
+      expect(obj.path[0].getCode()).to.be.equal('USD');
+      expect(obj.path[0].getIssuer()).to.be.equal(
+        'GBBM6BKZPEHWYO3E3YKREDPQXMS4VK35YLNU7NFBRI26RAN7GI5POFBB'
+      );
+      expect(obj.path[1].getCode()).to.be.equal('EUR');
+      expect(obj.path[1].getIssuer()).to.be.equal(
+        'GDTNXRLOJD2YEBPKK7KCMR7J33AAG5VZXHAJTHIG736D6LVEFLLLKPDL'
+      );
+    });
+
+    it('fails to create path payment operation with an invalid destination address', function() {
+      let opts = {
+        destination: 'GCEZW',
+        sendAmount: '20',
+        destMin: '50',
+        sendAsset: StellarBase.Asset.native(),
+        destAsset: new StellarBase.Asset(
+          'USD',
+          'GDGU5OAPHNPU5UCLE5RDJHG7PXZFQYWKCFOEXSXNMR6KRQRI5T6XXCD7'
+        )
+      };
+      expect(() => StellarBase.Operation.pathPaymentStrictSend(opts)).to.throw(
+        /destination is invalid/
+      );
+    });
+
+    it('fails to create path payment operation with an invalid sendAmount', function() {
+      let opts = {
+        destination: 'GCEZWKCA5VLDNRLN3RPRJMRZOX3Z6G5CHCGSNFHEYVXM3XOJMDS674JZ',
+        sendAmount: 20,
+        destMin: '50',
+        sendAsset: StellarBase.Asset.native(),
+        destAsset: new StellarBase.Asset(
+          'USD',
+          'GDGU5OAPHNPU5UCLE5RDJHG7PXZFQYWKCFOEXSXNMR6KRQRI5T6XXCD7'
+        )
+      };
+      expect(() => StellarBase.Operation.pathPaymentStrictSend(opts)).to.throw(
+        /sendAmount argument must be of type String/
+      );
+    });
+
+    it('fails to create path payment operation with an invalid destMin', function() {
+      let opts = {
+        destination: 'GCEZWKCA5VLDNRLN3RPRJMRZOX3Z6G5CHCGSNFHEYVXM3XOJMDS674JZ',
+        sendAmount: '20',
+        destMin: 50,
+        sendAsset: StellarBase.Asset.native(),
+        destAsset: new StellarBase.Asset(
+          'USD',
+          'GDGU5OAPHNPU5UCLE5RDJHG7PXZFQYWKCFOEXSXNMR6KRQRI5T6XXCD7'
+        )
+      };
+      expect(() => StellarBase.Operation.pathPaymentStrictSend(opts)).to.throw(
+        /destMin argument must be of type String/
       );
     });
   });